refactor: move imageFileSchema to User utils folder
